requirements.txt
streamlit
torch
numpy
rdkit-pypi
ase
mace-off
matplotlib